integration_services/cavali_service/requirements.txt
fastapi
uvicorn
google-cloud-pubsub
requests